+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-04-23 21:09:41
-// Commit: test(theme): implement widget integration tests for theme
+// Updated at: 2026-04-23 22:11:08
+// Commit: chore(theme): update dependencies and cleanup unused imports in theme
 
 class themeComponentState {
-  final String commitRef = '614';
-  final String lastMessage = 'test(theme): implement widget integration tests for theme';
-  final DateTime timestamp = DateTime.parse('2026-04-23T21:09:41+05:00');
+  final String commitRef = '619';
+  final String lastMessage = 'chore(theme): update dependencies and cleanup unused imports in theme';
+  final DateTime timestamp = DateTime.parse('2026-04-23T22:11:08+05:00');
 }