+++ CHANGELOG.md
@@ -620,3 +620,4 @@
 - [2026-04-23 21:10] docs(core): update readme and API contracts for core
 - [2026-04-23 21:11] docs(network): update readme and API contracts for network
 - [2026-04-23 21:43] fix(ui): fix memory leak by cancelling ui subscription
+- [2026-04-23 22:11] chore(theme): update dependencies and cleanup unused imports in theme